+++ src/test/java/org/sbrubbles/genericcons/CheckOnIterableTest.java
@@ -10,44 +10,44 @@ import java.util.Arrays;
 
 import org.junit.Test;
 
-public class CheckOnListTest {
+public class CheckOnIterableTest {
   @Test(expected = IllegalArgumentException.class)
   public void nullObjectsArgumentWithTypeArray() {
-    C.check().onList(null);
+    C.check().onIterable(null);
   }
   
   @Test(expected = IllegalArgumentException.class)
   public void nullObjectsArgumentWithListOfTypes() {
-    C.check(new ArrayList<Type>()).onList(null);
+    C.check(new ArrayList<Type>()).onIterable(null);
   }
   
   @Test
   public void emptyTypeArrayAgainstEmptyObjectList() {
-    assertTrue(C.check().onList(new ArrayList<Object>()));
+    assertTrue(C.check().onIterable(new ArrayList<Object>()));
   }
   
   @Test
   public void emptyTypeListAgainstEmptyObjectList() {
-    assertTrue(C.check(new ArrayList<Type>()).onList(new ArrayList<Object>()));
+    assertTrue(C.check(new ArrayList<Type>()).onIterable(new ArrayList<Object>()));
   }
   
   @SuppressWarnings("unchecked")
   @Test
   public void listSizeMismatch() {
-    assertFalse(C.check(Arrays.asList(String.class)).onList(Arrays.asList("", "")));
+    assertFalse(C.check(Arrays.asList(String.class)).onIterable(Arrays.asList("", "")));
   }
   
   @SuppressWarnings("unchecked")
   @Test
   public void sameTypeSingleArgumentMatch() {
-    assertTrue(C.check(Arrays.asList(String.class)).onList(Arrays.asList("")));
+    assertTrue(C.check(Arrays.asList(String.class)).onIterable(Arrays.asList("")));
   }
   
   @SuppressWarnings("unchecked")
   @Test
   public void sameTypesMultipleArgumentsMatch() {
     assertTrue(C.check(
-        Arrays.asList(String.class, Double.class, Object.class)).onList(
+        Arrays.asList(String.class, Double.class, Object.class)).onIterable(
         Arrays.asList("", new Double(1.0), new Object())));
   }
   
@@ -55,7 +55,7 @@ public class CheckOnListTest {
   @Test
   public void nullObjectsMultipleArgumentsMatch() {
     assertTrue(C.check(
-        Arrays.asList(String.class, Double.class, Object.class)).onList(
+        Arrays.asList(String.class, Double.class, Object.class)).onIterable(
         Arrays.asList(null, new Double(1.0), null)));
   }
   
@@ -63,7 +63,7 @@ public class CheckOnListTest {
   @Test
   public void nullTypesMultipleArgumentsMatch() {
     assertFalse(C.check(
-        Arrays.asList(String.class, null, Object.class)).onList(
+        Arrays.asList(String.class, null, Object.class)).onIterable(
         Arrays.asList(null, new Double(1.0), null)));
   }
   
@@ -71,7 +71,7 @@ public class CheckOnListTest {
   @Test
   public void subtypeSingleArgumentMatch() {
     assertTrue(C.check(
-        Arrays.asList(Serializable.class)).onList(
+        Arrays.asList(Serializable.class)).onIterable(
         Arrays.asList(new ArrayList<Object>())));
   }
   
@@ -79,7 +79,7 @@ public class CheckOnListTest {
   @Test
   public void subtypeSingleArgumentMismatch() {
     assertFalse(C.check(
-        Arrays.asList(Serializable.class)).onList(
+        Arrays.asList(Serializable.class)).onIterable(
         Arrays.asList(new Object())));
   }
   
@@ -87,7 +87,7 @@ public class CheckOnListTest {
   @Test
   public void subtypeMultipleArgumentsMatch() {
     assertTrue(C.check(
-        Arrays.asList(Serializable.class, Object.class, Number.class)).onList(
+        Arrays.asList(Serializable.class, Object.class, Number.class)).onIterable(
         Arrays.asList(new ArrayList<Object>(), "", 1.0)));
   }
   
@@ -95,7 +95,7 @@ public class CheckOnListTest {
   @Test
   public void subtypeMultipleArgumentsMismatch() {
     assertFalse(C.check(
-        Arrays.asList(Serializable.class, Object.class, Number.class)).onList(
+        Arrays.asList(Serializable.class, Object.class, Number.class)).onIterable(
         Arrays.asList(new ArrayList<Object>(), "", false)));
   }
   
@@ -103,7 +103,7 @@ public class CheckOnListTest {
   @Test
   public void tooManyArgumentsMismatch() {
     assertFalse(C.check(
-        Arrays.asList(Serializable.class, Object.class, Boolean.class)).onList(
+        Arrays.asList(Serializable.class, Object.class, Boolean.class)).onIterable(
         Arrays.asList(new ArrayList<Object>(), "", false, 1.0)));
   }
   
@@ -111,7 +111,7 @@ public class CheckOnListTest {
   @Test
   public void tooFewArgumentsMismatch() {
     assertFalse(C.check(
-        Arrays.asList(Serializable.class, Object.class, Boolean.class)).onList(
+        Arrays.asList(Serializable.class, Object.class, Boolean.class)).onIterable(
         Arrays.asList(new ArrayList<Object>(), "")));
   }
 }